refactor(db): move sqlc queries and generated code to internal/db
