+++ internal/db/queries/auth_requests.sql
