+++ internal/db/queries/cleanup.sql
