+++ internal/db/queries/device_codes.sql
