+++ internal/db/queries/sessions.sql
