+++ internal/db/queries/users.sql
