+++ internal/db/storeq/auth_requests.sql.go
