+++ internal/storage/storage_auth_tokens.go
@@ -9,7 +9,7 @@ import (
 	"github.com/zitadel/oidc/v3/pkg/oidc"
 	"github.com/zitadel/oidc/v3/pkg/op"
 
-	"github.com/kangheeyong/authgate/internal/storage/storeq"
+	"github.com/kangheeyong/authgate/internal/db/storeq"
 )
 
 // --- op.Storage: AuthStorage ---
